# app/BE/swagger/build.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Kids Keeper API Docs
  description: Kids Keeper API 문서
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - description: dev servers
    url: 'http://localhost:3000/api/'
paths:
  /districts:
    get:
      tags:
        - district
      summary: get all districts
      description: null
      parameters:
        - name: parent_code
          in: query
          required: false
          description: 상위 행정구역 코드
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    district_code:
                      type: string
                    district_name:
                      type: string
                    parent_code:
                      type: string
                    count:
                      type: string
  '/districts/{district_code}':
    get:
      tags:
        - district
      summary: get district by district code
      description: null
      parameters:
        - name: district_code
          in: path
          required: true
          description: 행정구역 코드
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  district_code:
                    type: string
                  district_name:
                    type: string
                  parent_code:
                    type: string
                  count:
                    type: string
  '/districts/{district_name}':
    get:
      tags:
        - district
      summary: get district by district name
      description: null
      parameters:
        - name: district_name
          in: path
          required: true
          description: 행정구역 이름
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  district_code:
                    type: string
                  district_name:
                    type: string
                  parent_code:
                    type: string
                  count:
                    type: string
  /centers:
    get:
      tags:
        - children care center
      summary: get all centers
      description: null
      parameters:
        - name: district_code
          in: query
          required: false
          description: 행정구역 코드
        - name: district_name
          in: query
          required: false
          description: 행정구역 이름
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    center_id:
                      type: integer
                    center_name:
                      type: string
                    operation_type:
                      type: string
                    operation_status:
                      type: string
                    zipcode:
                      type: string
                    address:
                      type: string
                    center_phone:
                      type: string
                    fax:
                      type: string
                    web_page:
                      type: string
                    latitude:
                      type: string
                    longtitude:
                      type: string
                    district_code:
                      type: string
                    district_name:
                      type: string
  '/centers/{center_id}':
    get:
      tags:
        - children care center
      summary: get center by center id
      description: null
      parameters:
        - name: center_id
          in: path
          required: true
          description: 어린이집 ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  center_id:
                    type: integer
                  center_name:
                    type: string
                  operation_type:
                    type: string
                  operation_status:
                    type: string
                  zipcode:
                    type: string
                  address:
                    type: string
                  center_phone:
                    type: string
                  fax:
                    type: string
                  web_page:
                    type: string
                  latitude:
                    type: string
                  longtitude:
                    type: string
                  district_code:
                    type: string
                  district_name:
                    type: string
  '/centers/{center_name}':
    get:
      tags:
        - children care center
      summary: get center by name
      description: null
      parameters:
        - name: center_name
          in: path
          required: true
          description: 어린이집 이름
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  center_id:
                    type: integer
                  center_name:
                    type: string
                  operation_type:
                    type: string
                  operation_status:
                    type: string
                  zipcode:
                    type: string
                  address:
                    type: string
                  center_phone:
                    type: string
                  fax:
                    type: string
                  web_page:
                    type: string
                  latitude:
                    type: string
                  longtitude:
                    type: string
                  district_code:
                    type: string
                  district_name:
                    type: string
  /cctvs:
    get:
      tags:
        - cctv
      summary: get all cctvs
      description: null
      parameters:
        - name: center_id
          in: query
          required: false
          description: 어린이집 id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    cctv_id:
                      type: integer
                    cctv_name:
                      type: string
                    cctv_mac:
                      type: string
                    quaity:
                      type: string
                    install_date:
                      type: string
                      format: date
                    uninstall_date:
                      type: string
                      format: date
                    center_id:
                      type: integer
                    center_name:
                      type: string
                    address:
                      type: string
    post:
      tags:
        - cctv
      summary: create cctv
      description: null
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                center_id:
                  type: integer
                cctv_name:
                  type: string
                cctv_mac:
                  type: string
                install_date:
                  type: string
                  format: date
                uninstall_date:
                  type: string
                  format: date
                quality:
                  type: string
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                type: object
                properties:
                  cctv_id:
                    type: integer
                  cctv_name:
                    type: string
                  cctv_mac:
                    type: string
                  install_date:
                    type: string
                    format: date
                  uninstall_date:
                    type: string
                    format: date
                  quality:
                    type: string
                  center_id:
                    type: integer
  '/cctvs/{cctv_mac}':
    put:
      tags:
        - cctv
      summary: update cctv by cctv_mac
      description: null
      parameters:
        - name: cctv_mac
          in: path
          required: true
          description: CCTV MAC 주소
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cctv_name:
                  type: string
                install_date:
                  type: string
                  format: date
                uninstall_date:
                  type: string
                  format: date
                quality:
                  type: string
      responses:
        '200':
          description: UPDATED
          content:
            application/json:
              schema:
                type: object
                properties:
                  cctv_id:
                    type: integer
                  cctv_name:
                    type: string
                  install_date:
                    type: string
                    format: date
                  uninstall_date:
                    type: string
                    format: date
                  quality:
                    type: string
                  center_id:
                    type: integer
    delete:
      tags:
        - cctv
      summary: delete cctv by cctv_mac
      description: null
      parameters:
        - name: cctv_mac
          in: path
          required: true
          description: CCTV MAC 주소
      responses:
        '204':
          description: DELETE
  /anomalies:
    get:
      tags:
        - anomaly
      summary: get all anomalies
      description: null
      parameters:
        - name: center_id
          in: query
          required: false
          description: 어린이집 id
        - name: center_name
          in: query
          required: false
          description: 어린이집 이름
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    anomaly_id:
                      type: integer
                    anomaly_type:
                      type: string
                    start_time:
                      type: string
                      format: date-time
                    end_time:
                      type: string
                      format: date-time
                    follow_up:
                      type: string
                    video_id:
                      type: integer
tags:
  - name: district
  - name: children care center
  - name: cctv
  - name: anomaly